% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basket.r
\name{basket}
\alias{basket}
\title{Fit Basket Trial Data}
\usage{
basket(formula, data, p0, method = mem_empirical_bayes, ...)
}
\arguments{
\item{formula}{the number of responders, enrollees, and possibly basket
names for a basket trial.}

\item{data}{the basket trial data}

\item{p0}{the null (or failure probability) for the basket trial.}

\item{method}{the method for fitting the basket trial. This could be
mem_empirical_bayes, mem_full_bayes, sem, or a custom fitting function.}

\item{...}{other parameters to be passed to the supplied
\code{method} argument.}
}
\description{
TODO FINISH THIS
}
\examples{
# WRITE THIS
# fit <- basket(responders ~ evaluable | basket, data = vemu_wide)
}
