% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{basket_map}
\alias{basket_map}
\title{Get the Basketwise Maximum A Posteriori Probability Matrix}
\usage{
basket_map(x)
}
\arguments{
\item{x}{either an exchangeability model or basket object.}
}
\description{
TODO: WRITE THIS
}
\examples{
# TODO: WRITE THIS
}
