% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{basket_pep}
\alias{basket_pep}
\title{Get the Basketwise Posterior Exchangeability Matrix}
\usage{
basket_pep(x)
}
\arguments{
\item{x}{either an exchangeability model or basket object.}
}
\description{
TODO: WRITE THIS
}
\examples{
# TODO: WRITE THIS
}
