% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{cluster_arms}
\alias{cluster_arms}
\title{Get the Members of each cluster}
\usage{
cluster_arms(x)
}
\arguments{
\item{x}{either an exchangeability model or basket object.}
}
\description{
TODO: WRITE THIS
}
\examples{
# TODO: WRITE THIS
}
