% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{cluster_map}
\alias{cluster_map}
\title{Get the Clusterwise Maximum A Posteriori Probability Matrix}
\usage{
cluster_map(x)
}
\arguments{
\item{x}{either an exchangeability model or basket object.}
}
\description{
TODO: WRITE THIS
}
\examples{
# TODO: WRITE THIS
}
