% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basket.r
\name{cluster_mean}
\alias{cluster_mean}
\title{Calculate the Exchangeability Cluster Means}
\usage{
cluster_mean(x, method = "maximizer")
}
\arguments{
\item{x}{the exchangeability model}

\item{method}{the exchangeability model to get the clusters for. This
can be either "maximzer" or "pep" if a model includes a posterior 
exchangeability probability.}
}
\description{
TODO FINISH THIS
}
\examples{
# WRITE THIS
}
