% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{cluster_pep}
\alias{cluster_pep}
\title{Get the Clusterwise Posterior Exchangeability Matrix}
\usage{
cluster_pep(x)
}
\arguments{
\item{x}{either an exchangeability model or basket object.}
}
\description{
TODO: WRITE THIS
}
\examples{
# TODO: WRITE THIS
}
