% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basket.r
\name{get_seed}
\alias{get_seed}
\title{Get the Current Random Seed}
\usage{
get_seed()
}
\description{
Get the value of the current random seed. If one has not been
initialized, then initialize it and return the new seed value.
}
