% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kbeta.r
\name{kbeta}
\alias{kbeta}
\alias{dkbeta}
\alias{pkbeta}
\alias{qkbeta}
\alias{rkbeta}
\title{The K-Beta Disribution}
\usage{
dkbeta(x, shape1, shape2, a = -1, b = 1, log = FALSE)
pkbeta(q, shape1, shape2, a = -1, b = 1, lower.tail = TRUE, log.p = FALSE)
qkbeta(p, shape1, shape2, a = -1, b = 1, lower.tail = TRUE, log.p = FALSE)
rkbeta(n, shape1, shape2, a = -1, b = 1)
}
\arguments{
\item{x}{vector of values in the support.}

\item{shape1}{the first non-negative parameters of the Beta distribution.}

\item{shape2}{the second non-negative parameters of the Beta distribution.}

\item{a}{the lower bound of the support of the distribution (default -1).}

\item{b}{the upper bound of the support of the distribution (default 1).}

\item{log}{should the distribution value be shown on the log scale?
(default FALSE)}

\item{q}{vector of quantiles.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If ‘length(n) > 1’, the length is
taken to be the number required.}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x],
otherwise, P[X > x].}
}
\description{
Density, distribution function, quantile function, and
random geration for the K-Beta distribution with shape parameters
\code{shape1}, \code{shape2}, \code{a}, and \code{b}. The shape 
parameters are the same as those of the beta distribution. The
\code{a} parameter defines the smallest value in the support of the 
distribution and the \code{b} parameter defines the largest.
}
\examples{
library(ggplot2)
a <- 1
b <- 3
shape1 <- 0.5
shape2 <- 0.5
x <- seq(a, b, by = 0.01)
d <- data.frame(x = x, y = dkbeta(x, 0.5, 0.5, a, b))
ggplot(d, aes(x = x, y = y)) + 
  geom_area(fill = "black", alpha = 0.7) +
  xlim(0, 3)
}
