% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mem-empirical-bayes.r
\name{mem_empirical_bayes}
\alias{mem_empirical_bayes}
\title{MEM Empirical Bayes}
\usage{
mem_empirical_bayes(responses, size, name = NULL, p0 = 0.25,
  shape1 = 0.5, shape2 = 0.5, hpd_alpha = 0.05, upper_bound = 1,
  lower_bound = 0, call = NULL)
}
\arguments{
\item{responses}{the number of responses in each basket.}

\item{size}{the size of each basket.}

\item{name}{the name of each basket (default: NULL - no basket names).}

\item{p0}{the null response rate for the poster probability calculation
(default 0.25).}

\item{shape1}{the first shape parameter(s) for the prior of each basket
(default 0.5).}

\item{shape2}{the second shape parameter(s) for the prior of each basket
(default 0.5).}

\item{hpd_alpha}{the highest posterior density trial significance.}

\item{upper_bound}{for constrained empirical Bayes, the upper bound on the
inclusion probability (default 1).}

\item{lower_bound}{for constrained empirical Bayes, the lower bound on the
inclusion probability (default 0).}

\item{call}{the call of the function (default NULL).}
}
\description{
Fit the MEM model using empirical Bayesian inference.
}
\examples{
# 5 baskets, each with enrollement size 5
trial_sizes <- rep(5, 5)

# The response rates for the baskets.
resp_rate <- 0.15

# The trials: a column of the number of responses and a column of the
# the size of each trial.
trials <- data.frame(
  responses = rbinom(trial_sizes, trial_sizes, resp_rate),
  size = trial_sizes
)

mem_empirical_bayes(trials$responses, trials$size)
}
