% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mem-mcmc.r
\name{mem_mcmc}
\alias{mem_mcmc}
\title{Fit the MEM Model using MCMC}
\usage{
mem_mcmc(responses, size, name, p0 = 0.15, shape1 = 0.5,
  shape2 = 0.5, prior = diag(length(responses))/2 + matrix(0.5, nrow =
  length(responses), ncol = length(responses)), hpd_alpha = 0.05,
  alternative = "greater", mcmc_iter = 10000, Initial = round(prior - 0.001),
  seed = 1000, call = NULL)
}
\arguments{
\item{responses}{the number of responses in each subgroup.}

\item{size}{the size of each subgroup.}

\item{name}{the name of each subgroup.}

\item{p0}{the null response rate for the poster probability calculation
(default 0.15).}

\item{shape1}{the first beta shape parameter(s) for the prior distribution of the response probability for each basket
(default 0.5).}

\item{shape2}{the second beta shape parameter(s) for the prior distribution of the response probability for each basket
(default 0.5).}

\item{prior}{the symemtric matrix giving the prior exchangeability probabilities
for each pair of baskets. The default is on on the main diagonal and 0.5
elsewhere.}

\item{hpd_alpha}{the highest posterior density interval significance threshold.}

\item{alternative}{the alternative case definition (default greater)}

\item{mcmc_iter}{the number of MCMC iterations.}

\item{Initial}{the initial MEM matrix used to start MCMC sampling (default is the diagonal matrix).}

\item{seed}{the random number seed.}

\item{call}{the call of the function.}
}
\description{
Fit the MEM model using full Bayesian posterior inference using MCMC sampling using the Metropolis-Hastings algorithm.
}
\examples{
# 3 subgroups, each with enrollement size 5
trial_sizes <- rep(5, 3)

# The response rates for the subgroups.
resp_rate <- 0.15

# The trial data: a column of the number of responses and a column of the
# the size of each subgroup.
trials <- data.frame(
  responses = rbinom(trial_sizes, trial_sizes, resp_rate),
  size = trial_sizes,
  name = letters[1:3]
)
res <- mem_mcmc(trials$responses, trials$size)
}
