% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot_all_exchangeability}
\alias{plot_all_exchangeability}
\title{Plot the Prior, MAP, and PEP of a Basket Trial}
\usage{
plot_all_exchangeability(x, plotList, ...)
}
\arguments{
\item{x}{the exchangeability model.}

\item{plotList}{TODO: WHAT IS THIS?}

\item{...}{other options. See Details for more information.}
}
\description{
: TODO: WRITE THIS
}
