% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot_density}
\alias{plot_density}
\title{Plot the Posterior Distributions of Response Probabilities for each Subgroup}
\usage{
plot_density(x, ...)
}
\arguments{
\item{x}{the MEM model.}

\item{...}{other options. See Details for more information.}
}
\description{
Plot the density of MEM model.
}
\details{
Plot the density of MEM model.
}
\examples{
\dontrun{
# Create an MEM analysis of the Vemurafenib trial data.

data(vemu_wide)

mem_analysis <- mem_exact(vemu_wide$responders,
                          vemu_wide$evaluable,
                          vemu_wide$baskets)

plot_density(mem_analysis$basket)
}}
