% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot_pep}
\alias{plot_pep}
\title{Plot the Posterior Exchangeability of a Basket Trial}
\usage{
plot_pep(x, ...)
}
\arguments{
\item{x}{the exchangeability model.}

\item{...}{other options. See Details for more information.}
}
\description{
TODO: WRITE THIS
}
\details{
TODO: WRITE THIS
}
\examples{
# WRITE THIS
}
