% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.r
\docType{data}
\name{eb_reference}
\alias{eb_reference}
\alias{eb_reference_ub}
\alias{fb_reference}
\title{Reference Outputs for Testing}
\description{
There are three reference data sets used for testing:
eb_reference, eb_reference ub, and fb_reference. These are the the
outputs generated by the reference code (in inst/code-from-brian) and
are used to detect changes in the output of our implementation.
}
