% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem-model.r
\name{sem}
\alias{sem}
\title{Single-Source Model}
\usage{
sem(responses, size, name, p0 = 0.25, mu = 0, M = 0, inv_sig = 1,
  num_chains = 3, quiet = TRUE, num_iterations = 3000,
  progress_bar = interactive())
}
\description{
Fit the single source model.
}
\references{
Berry SM, Broglio KR, Groshen S, Berry DA. Bayesian 
hierarchical modeling of patient subpopulations: efficient designs of 
phase II oncology clinical trials. Clinical Trials. 2013 Oct;10(5):720-34.
}
