% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress.R
\name{stress_test}
\alias{stress_test}
\title{A simple stress test}
\usage{
stress_test(
  email = "statprep.annie@gmail.com",
  key = "1jiV11I9tlEfAhRrkYqn6kPtkNma04TEpSCk-XMSuIRM"
)
}
\description{
A simple stress test
}
