% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.r
\name{update_result}
\alias{update_result}
\title{Update Full Bayes results with different p0 values}
\usage{
update_result(res, p0 = 0.15, alternative = "greater")
}
\arguments{
\item{res}{TODO: WHAT IS THIS?}

\item{p0}{the null response rate for the poster probability calculation
(default 0.15).}

\item{alternative}{the alternative case defination (default greater)}
}
\description{
Update Full Bayes results with different p0 values and alternative
}
\examples{
# MHResult1New <- update_result(MHResult1, 0.25)
}
