% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.r
\docType{data}
\name{vemu}
\alias{vemu}
\alias{vemu_wide}
\title{Summarize Data from the Vemurafinib Study}
\description{
The `vemu` and `vemu_wide` data sets provides trial data for six subgroups described in article ``Vemurafenib in multiple nonmelanoma cancers with braf v600
mutations''. A total of 18 responders were observed among the
84 patients contributing evaluable outcomes for statistical estimation.
Observed response rates varied from $42\%$ and $43\%$ for subgroups of NSCLC
and ECD or LCH to 0 and $4\%,$ for CRC with vemurafenib mono and
combination therapies, respectively. With two responders of seven patients, ATC
was associated with a $29\%$ response rate, while one responder of eight
patients was observed in the cholangiocarcinoma subgroup. Contrasting
favorable results for preliminary vemurafenib activity among NSCLC and ECD
or LCH patients with less favorable results for CRC patients, the authors
concluded that nonmelanoma tumor types harboring $BRAF^{V600}$ mutations
failed to respond uniformly to BRAF-targeted therapy giving credence to
more conventional organ-specific nosology when compared to molecular tumor
nosology.

Later, in the ``Statistical challenges posed by basket trials: sensitivity
analysis of the Vemurafenib study'' it was shown that patient-enrollment
heterogeneity among the subgroups may explain in part the discordant response rates among clinical indications.
}
\references{
Hyman DM, Puzanov I, Subbiah V, Faris JE, Chau I, Blay JY,
Wolf J, Raje NS, Diamond EL, Hollebecque A, et al. Vemurafenib in multiple
nonmelanoma cancers with braf v600 mutations. New England Journal of
Medicine 2015; \strong{373}(8):726
}
