% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{basket_map}
\alias{basket_map}
\title{Return the Basketwise Maximum A Posteriori Probability Matrix}
\usage{
basket_map(x)
}
\arguments{
\item{x}{either an MEM model or basket object.}
}
\description{
Analyses compute the maximum a posterior (MAP) MEM matrix which describes the single set of pairwise exchangeability relationships best supported by the posterior. This function returns returns a symmetric matrix of zeros and ones describing the MAP MEM.
}
\examples{
# Create an MEM analysis of the Vemurafenib trial data.

data(vemu_wide)

mem_analysis <- mem_exact(vemu_wide$responders, 
                          vemu_wide$evaluable,
                          vemu_wide$baskets)

# Get basket MAPs.
basket_map(mem_analysis)
}
