% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{basket_pep}
\alias{basket_pep}
\title{Return the Basketwise Posterior Exchangeability Matrix}
\usage{
basket_pep(x)
}
\arguments{
\item{x}{either an MEM model or basket object.}
}
\description{
MEM analyses include the posterior exachangeability 
probabilities (PEP) of included subgroups describing the probability
that any two arms are exchangeable. This function returns the matrix
of those probabilities.
}
\examples{
\dontrun{
# Create an MEM analysis of the Vemurafenib trial data.

data(vemu_wide)

mem_analysis <- mem_exact(vemu_wide$responders, 
                          vemu_wide$evaluable,
                          vemu_wide$baskets)

# Get the PEP for baskets.
basket_pep(mem_analysis)
}}
