% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{cluster_arms}
\alias{cluster_arms}
\title{Get the Cluster Members of MEM Models}
\usage{
cluster_arms(x)
}
\arguments{
\item{x}{either an exchangeability model or basket object.}
}
\value{
A named list is returned where the name is the cluster name and
each element of the list is comprise do a character vector of the 
baskets in each cluster.
}
\description{
Object returned by the `mem_mcmc()` and `mem_exact()` include
information about the arms in the trials and the cluster 
composed of sets of similar arms. This function returns the name
of each arm in a cluster.
}
\examples{
# Create an MEM analysis of the first three arms in the Vemurafenib
# trial data.
data(vemu_wide)

mem_analysis <- mem_exact(vemu_wide$responders[1:3], 
                          vemu_wide$evaluable[1:3],
                          vemu_wide$baskets[1:3])

# Get the arms in the clusters.
cluster_arms(mem_analysis)
}
