% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{cluster_map}
\alias{cluster_map}
\title{Return the Clusterwise Maximum A Posteriori Probability Matrix}
\usage{
cluster_map(x)
}
\arguments{
\item{x}{either an MEM model or basket object.}
}
\description{
Analyses compute the maximum a posterior (MAP) MEM matrix which describes the single set of pairwise exchangeability relationships best supported by the posterior. This function returns the matrix
of those relationships after clustering.
}
\examples{
\dontrun{
# Create an MEM analysis of the Vemurafenib trial data.

data(vemu_wide)

mem_analysis <- mem_exact(vemu_wide$responders, 
                          vemu_wide$evaluable,
                          vemu_wide$baskets)

# Get the cluster MAPs.
cluster_map(mem_analysis)
}}
