% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-pep.r
\name{cluster_PEP}
\alias{cluster_PEP}
\title{Cluster results based on PEP}
\usage{
cluster_PEP(responses, size, name, models, pweights, PEP, p0 = 0.15,
  shape1 = 0.5, shape2 = 0.5, HPD.alpha = 0.05,
  alternative = "greater", seed = 1000, num_samples = 20000,
  call = NULL)
}
\arguments{
\item{responses}{the number of responses in each basket.}

\item{size}{the size of each basket.}

\item{name}{the name of each basket.}

\item{models}{the models calculated from the MEM computation.}

\item{pweights}{the weights calculated from the MEM computation.}

\item{PEP}{the PEP matrix calculated from the MEM computation.}

\item{p0}{the null response rate for the poster probability calculation
(default 0.15).}

\item{shape1}{the first shape parameter(s) for the prior of each basket
(default 0.5).}

\item{shape2}{the second shape parameter(s) for the prior of each basket
(default 0.5).}

\item{HPD.alpha}{the highest posterior density trial significance.}

\item{alternative}{the alternative case definition (default "greater"). 
TODO: WHAT DOES THIS MEAN? WHAT ARE THE ALTERNATIVES?}

\item{seed}{the random number seed.}

\item{num_samples}{the number of sample iterations for each cluster.}

\item{call}{the call of the function.}
}
\description{
Conductthe the basket clustering based on the PEP information and provide the information on the cluster level. .
}
\examples{
# 5 baskets, each with enrollement size 5
trial_sizes <- rep(15, 6)
# The response rates for the baskets.
resp_rate <- 0.15
# The trials: a column of the number of responses, a column of the
# the size of each trial.
trials <- data.frame(
  responses = c(1, 4, 5, 0, 1, 6),
  size = trial_sizes
)
MHResult2 <- mem_full_bayes_mcmc(trials$responses, trials$size,
  name = c(" D1 ", " D2 ", " D3 ", " D4 ", " D5 ", " D6 ")
)

#result <- cluster_PEP(
#  responses = trials$responses, size = trials$size,
#  name = c(" D1 ", " D2 ", " D3 ", " D4 ", " D5 ", " D6 "),
#  models = MHResult2$models, pweights = MHResult2$pweights, p0 = 0.15,
#  PEP = MHResult2$PEP
#)
print(trials)
#print(result$clusters)
#print(result$HPD)
#print(result$mean_est)
}
