% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kbeta.r
\name{kbeta}
\alias{kbeta}
\alias{dkbeta}
\alias{pkbeta}
\alias{qkbeta}
\alias{rkbeta}
\title{The K-Beta Disribution}
\usage{
dkbeta(x, shape1, shape2, a = 0, b = 1, log = FALSE)
pkbeta(q, shape1, shape2, a = 0, b = 1, lower.tail = TRUE, log.p = FALSE)
qkbeta(p, shape1, shape2, a = 0, b = 0, lower.tail = TRUE, log.p = FALSE)
rkbeta(n, shape1, shape2, a = 0, b = 0)
}
\description{
Density, distribution function, quantile function, and
random geration for the K-Beta distribution with shape parameters
\code{shape1}, \code{shape2}, \code{a}, and \code{b}. The shape 
parameters are the same as those of the beta distribution. The
\code{a} parameter defines the smallest value in the support of the 
distribution and the \code{b} parameter defines the largest.
}
\examples{
a <- 1
b <- 3
shape1 <- 0.5
shape2 <- 0.5
x <- seq(a, b, by = 0.01)
d <- tibble(x = x, y = dkbeta(x, 0.5, 0.5, a, b))
ggplot(d, aes(x = x, y = y)) + 
  geom_area(fill = "black", alpha = 0.7) +
  xlim(0, 3)
}
