% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mem-full-bayes.r
\name{mem_full_bayes}
\alias{mem_full_bayes}
\title{MEM Full Bayes}
\usage{
mem_full_bayes(responses, size, name, p0 = 0.25, shape1 = 0.5,
  shape2 = 0.5, prior_inclusion = diag(length(responses))/2 +
  matrix(0.5, nrow = length(responses), ncol = length(responses)),
  hpd_alpha = 0.05, call = NULL)
}
\arguments{
\item{responses}{the number of responses in each basket.}

\item{size}{the size of each basket.}

\item{name}{the name of each basket.}

\item{p0}{the null response rate for the poster probability calculation
(default 0.25).}

\item{shape1}{the first shape parameter(s) for the prior of each basket
(default 0.5).}

\item{shape2}{the second shape parameter(s) for the prior of each basket
(default 0.5).}

\item{prior_inclusion}{the matrix giving the prior inclusion probability
for each pair of baskets. The default is on on the main diagonal and 0.5
elsewhere.}

\item{hpd_alpha}{the highest posterior density trial significance.}

\item{call}{the call of the function (default NULL).}
}
\description{
Fit the MEM model using full Bayesian inference.
}
\examples{
# 5 baskets, each with enrollement size 5
trial_sizes <- rep(5, 5) 

# The response rates for the baskets.
resp_rate <- 0.15

# The trials: a column of the number of responses and a column of the
# the size of each trial.
trials <- data.frame(responses=rbinom(trial_sizes, trial_sizes, resp_rate), 
                     size = trial_sizes)

mem_full_bayes(trials$responses, trials$size)
}
