% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot_density}
\alias{plot_density}
\title{Plot the Densities of Baskets in a Trial}
\usage{
plot_density(x, ...)
}
\arguments{
\item{x}{the exchangeability model.}

\item{...}{other options. See Details for more information.}
}
\description{
TODO: WRITE THIS
}
\details{
TODO WRITE THIS TALK ABOUT ... OPTIONS
}
\examples{
\dontrun{
# Create an MEM analysis of the first three arms in the Vemurafenib
# trial data.
data(vemu_wide)

mem_analysis <- mem_exact(vemu_wide$responders,
                          vemu_wide$evaluable,
                          vemu_wide$baskets)

plot_density(mem_analysis)
}}
