% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot_map}
\alias{plot_map}
\title{Plot the Map Exchangeability of a Basket Trial}
\usage{
plot_map(x, ...)
}
\arguments{
\item{x}{the exchangeability model.}

\item{...}{other options. See Details for more information.}
}
\description{
Plot the MAP of the exchaneability model.
}
\details{
TODO: WRITE THIS
}
\examples{
\dontrun{
# Create an MEM analysisthe Vemurafenib trial data.
data(vemu_wide)

mem_analysis <- mem_exact(vemu_wide$responders,
                          vemu_wide$evaluable,
                          vemu_wide$baskets)

plot_map(mem_analysis)
}}
