% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot_map}
\alias{plot_map}
\title{Plot the Map Exchangeability of a Basket Trial}
\usage{
plot_map(x, ...)
}
\arguments{
\item{x}{\code{basket} element of the exchangeability model.}

\item{...}{other options passed to ggplot2 to alter the visual}
}
\description{
The Maximum A Posteriori Probability (MAP) of an MEM is the
estimate of the exchangeability structure of a basket trial. This function
visualizes this matrix as an exchangeogram.
}
\details{

}
\examples{
\donttest{
# Create an MEM analysis of the Vemurafenib trial data.
data(vemu_wide)

mem_analysis <- mem_exact(vemu_wide$responders,
                          vemu_wide$evaluable,
                          vemu_wide$baskets)

plot_map(mem_analysis$basket)
}
}
