% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot_mem}
\alias{plot_mem}
\title{Plot the Prior, MAP, and PEP of a Basket Trial}
\usage{
plot_mem(x, type = c("prior", "map", "pep"), ...)
}
\arguments{
\item{x}{the exchangeability model.}

\item{type}{the plot type that will be plotted.}

\item{...}{other options. See Details for more information.}
}
\description{
Plot multiple sub-plot of the exchangeability model, including "prior", "map", and "pep".
}
\examples{
\dontrun{
# Create an MEM analysisthe Vemurafenib trial data.
data(vemu_wide)

mem_analysis <- mem_exact(vemu_wide$responders,
                          vemu_wide$evaluable,
                          vemu_wide$baskets)

plot_mem(mem_analysis)
}}