% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot_mem}
\alias{plot_mem}
\title{Plot the Prior, MAP, and PEP of a Basket Trial}
\usage{
plot_mem(x, plot_list, ...)
}
\arguments{
\item{x}{the exchangeability model.}

\item{plot_list}{TODO: WHAT IS THIS?}

\item{...}{other options. See Details for more information.}
}
\description{
: TODO: WRITE THIS
}
