% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basket.r
\name{sample_posterior}
\alias{sample_posterior}
\title{Sample Posterior Samples from a Basket Trial}
\usage{
sample_posterior(model, num_samples = 10000)
}
\arguments{
\item{model}{the exchangeability model}

\item{num_samples}{the number of samples to draw. Default 10000}
}
\description{
TODO FINISH THIS
}
\examples{
# WRITE THIS
}
