% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.r
\name{summary.full_bayes}
\alias{summary.full_bayes}
\title{Make the summary table}
\usage{
\method{summary}{full_bayes}(object, ...)
}
\arguments{
\item{object}{the full_bayes class object..}
}
\description{
From the input full_bayes class object, summarize the CDF, HPD, ESS, Mean, and Median objectults.
}
\examples{

}
